Move ImageWithMultipleVariants to utils
