+++ app/src/main/java/me/saket/dank/ui/submission/SubmissionImageLoader.java
@@ -25,7 +25,7 @@ import io.reactivex.Scheduler;
 import io.reactivex.Single;
 import me.saket.dank.R;
 import me.saket.dank.ui.preferences.NetworkStrategy;
-import me.saket.dank.ui.submission.adapter.ImageWithMultipleVariants;
+import me.saket.dank.utils.ImageWithMultipleVariants;
 import me.saket.dank.urlparser.MediaLink;
 import me.saket.dank.utils.NetworkStateListener;
 import me.saket.dank.utils.Optional;
