+++ app/src/main/java/me/saket/dank/ui/submission/adapter/SubmissionContentLinkUiConstructor.java
@@ -35,10 +35,7 @@ import me.saket.dank.data.ErrorResolver;
 import me.saket.dank.data.LinkMetadataRepository;
 import me.saket.dank.data.ResolvedError;
 import me.saket.dank.urlparser.*;
-import me.saket.dank.utils.Colors;
-import me.saket.dank.utils.Optional;
-import me.saket.dank.utils.Pair;
-import me.saket.dank.utils.Urls;
+import me.saket.dank.utils.*;
 import me.saket.dank.utils.glide.GlideCircularTransformation;
 import me.thanel.dawn.linkunfurler.LinkMetadata;
 import timber.log.Timber;
